Undo the attempt to turn zenixarch into qubesos
It was a cool idea, but I quickly realized this was not only a
monumental task that requires a LOT of knowledge, but also I feel like
this is leaning too hard into security at the cost of performance,
especially when I can't truly replicate Qubes' model.

The idea I had was to have a minimalist host just for graphics and being
the hypervisor, then I had an arch template that is shared across all
the VMs. The first VM is the user VM which was meant for user apps and
is the only one with /home access, then the rest of the VM's are the
network stack: proxy VM (wireguard, adguardhome) --> firewall VM
(nftables) --> network VM (iwd).

That all said, I realized I could integrate Whonix instead, which would
be very similar architecturally to my idea anyways.