Replace doas with su 2: electric boogaloo
Last time I didn't know pkexec existed which could be used as the pacman
auth for makepkg, and that was quite literally the only problem I had
with that whole change.

This time I'm gonna commit to su for unix reasons, which does mean that
having a password on the user account is completely useless because
never once in this setup is it typed.